Changed Weller WSP80 Rid1 from 120 ohms to 330 ohms
Value of Rid1 is changed from 120 to 330 ohms in order to reduce false identification of JBC C245 (Rid1=150 ohms, same Rid2) to this model, which results in a soldering cartridge becoming incandescent. The root cause of the false reading of Rid1 has not been identified. However, this breaking change seems desiderable from safety point of view.